refactor: 이미지 파일 이동
