+++ src/constants/categoryImage.ts
@@ -1,10 +1,10 @@
 import type { Category } from "../types/restaurant";
-import categoryKoreanImage from "../../templates/category-korean.png";
-import categoryAsianImage from "../../templates/category-asian.png";
-import categoryChineseImage from "../../templates/category-chinese.png";
-import categoryJapaneseImage from "../../templates/category-japanese.png";
-import categoryWesternImage from "../../templates/category-western.png";
-import categoryEtcImage from "../../templates/category-etc.png";
+import categoryKoreanImage from "../images/category/category-korean.png";
+import categoryAsianImage from "../images/category/category-asian.png";
+import categoryChineseImage from "../images/category/category-chinese.png";
+import categoryJapaneseImage from "../images/category/category-japanese.png";
+import categoryWesternImage from "../images/category/category-western.png";
+import categoryEtcImage from "../images/category/category-etc.png";
 
 const categoryImages: Record<Category, string> = {
   한식: categoryKoreanImage,
