+++ src/images/favorite-icon-filled.png
